% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_summary.R
\name{add_summary}
\alias{add_summary}
\title{clean variables and create summary stats}
\usage{
add_summary(ds1)
}
\arguments{
\item{ds1}{data.frame with variables in columns}
}
\value{
data.frame of summary stats that can be row-appended to the original data frame
}
\description{
clean variables and create summary stats
}
\details{
Variable type is inferred (str, num, bool, logic) and the following summary stats are created: n (number of entries), completeness (% not NA, null, na, nan, inf, empty strings), number of unique values, univariate descriptors (median and range for number variables, count and fraction for categorical variables)
}
\examples{
\dontrun{
query=dbSendQuery(ch,"SELECT * FROM app_data;")
ds1=dbFetch(query)
dbClearResult(query)
test=add_summary(ds1)
t1=as.data.frame(test)
WriteXLS("t1", "sums.xlsx", row.names = T, AdjWidth = F, BoldHeaderRow = T, FreezeRow = 1, FreezeCol = 1, AllText = T)

}
}
