% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_numeric.R
\name{df_to_numeric}
\alias{df_to_numeric}
\title{clean data and prepare for pca}
\usage{
df_to_numeric(t1, df1, imp_cutoff_perc = 25, un_filter_perc = 20)
}
\arguments{
\item{t1}{output uf add_summary fct}

\item{df1}{data.frame to be cleaned}

\item{imp_cutoff_perc}{percentage of missing values for a variable to be excluded}

\item{un_filter_perc}{percentage of unique values of a categorical variable to be excluded}
}
\value{
data.frame of cleaned numberic variables
}
\description{
data canversion to numeric, dummy variables for multi-level categorical vars, missing value imputation
}
