% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.clean_na}
\alias{.clean_na}
\title{Set non-informative values to NA}
\usage{
.clean_na(x)
}
\arguments{
\item{x}{variable to be cleaned}
}
\value{
variable with values such as "na" set to NA
}
\description{
Set non-informative values to NA
}
\details{
Replaced with NA are strings: NA, nan, inf, "NA", "NAN", "null"
}
