% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.comb_xy}
\alias{.comb_xy}
\title{Combine uni and/or multilevel variables, rm NA's}
\usage{
.comb_xy(x, y)
}
\arguments{
\item{x}{array, categorical or numeric variable to be merged with y}

\item{y}{array, categorical or numeric variable to be merged with x}
}
\value{
list, 1: merged x and y, 2: indices of NA in x or y
}
\description{
Combine uni and/or multilevel variables, rm NA's
}
