% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.combine_lowFreqLevels}
\alias{.combine_lowFreqLevels}
\title{Merge low freq levels into a single level}
\usage{
.combine_lowFreqLevels(x, n = 2, repl = "Other", includeLevN = T)
}
\arguments{
\item{x}{array, categorical variable to be processed}

\item{n}{num Percent frequency value of x level, for assigning new level}

\item{repl}{char, Name of new level}

\item{includeLevN}{logic, Level name indicates number of merges}
}
\value{
x with low freq levels merged into new level
}
\description{
Merge low freq levels into a single level
}
