% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.conv_dtype}
\alias{.conv_dtype}
\title{Variable class conversion}
\usage{
.conv_dtype(x, dtype = c("num", "str", "fac"), levs = NULL)
}
\arguments{
\item{x}{variable to be converted}

\item{dtype}{desired variable class (numeric, string, factor)}

\item{levs}{levels in case of class factor}
}
\value{
converted x
}
\description{
Variable class conversion
}
