% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.count_na}
\alias{.count_na}
\title{Counting the number of NA}
\usage{
.count_na(x, y)
}
\arguments{
\item{x}{variable for stratification}
}
\value{
df counting NA, abs, %
}
\description{
Counting the number of NA
}
