% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.countsFreq_chiSq}
\alias{.countsFreq_chiSq}
\title{Univariate stats for a categorical variable}
\usage{
.countsFreq_chiSq(ds, x.multi, y.multy)
}
\arguments{
\item{ds}{data.frame, categorical variable and segmentation variable}

\item{x.multi}{logic, is x cleaned multi-level variable?}

\item{y.multi}{logic, is y cleaned multi-level variable?}
}
\value{
list, 1: contingency table(x,y) and y, 2: Chi-squared p value
}
\description{
Univariate stats for a categorical variable
}
