% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.descrStats_x}
\alias{.descrStats_x}
\title{Descriptive stats for a numeric variable startified by second variable}
\usage{
.descrStats_x(out)
}
\arguments{
\item{out}{list generated with .summary_var}
}
\value{
array of quantiles (range, IQR point estimates, median), length and %total for each segmentation level
}
\description{
Descriptive stats for a numeric variable startified by second variable
}
