% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.num_descrStats}
\alias{.num_descrStats}
\title{Descriptive stats for a numeric variable startified by second variable}
\usage{
.num_descrStats(x, y)
}
\arguments{
\item{x}{num array, numeric variable}

\item{y}{cat array, segementation variable}
}
\value{
array of quantiles (range, IQR point estimates, median), length and %total for each segmentation level
}
\description{
Descriptive stats for a numeric variable startified by second variable
}
