% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.num_groupComp}
\alias{.num_groupComp}
\title{Statistical group comparison for a numeric variable}
\usage{
.num_groupComp(x, y)
}
\arguments{
\item{x}{num array, variable for group comparison}

\item{y}{cat array, segementation variable defining groups}
}
\value{
list, 1: p value for comparing all groups and y, 2: contingency table y
}
\description{
Statistical group comparison for a numeric variable
}
\details{
Pariwise group comparison is performed with fct .num_groupComp_pariwise
}
