% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.num_groupComp_pairwise}
\alias{.num_groupComp_pairwise}
\title{Pairwise statistical group comparison for a numeric variable}
\usage{
.num_groupComp_pairwise(x, y)
}
\arguments{
\item{x}{num array, variable for group comparison}

\item{y}{cat array, segementation variable defining groups}
}
\value{
matrix, row 1: p value, row 2: Cliff's delta effect size
}
\description{
Pairwise statistical group comparison for a numeric variable
}
