% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.split_multiresp}
\alias{.split_multiresp}
\title{Split multilevel variables}
\usage{
.split_multiresp(x, sep = ", ", return = c("list", "vector"))
}
\arguments{
\item{x}{Multilevel variable to be split, given as string}

\item{sep}{Symbol that separates levels within a string}

\item{return}{return list or vector (see Details)}
}
\value{
converted x
}
\description{
Split multilevel variables
}
\details{
return argument is list: each x element is list element, return argument is vector: list is vectorised to 1D array, length does not equal number of elements in x then.
}
