% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.split_to_dummy}
\alias{.split_to_dummy}
\title{split multilevel variable and create dummy}
\usage{
.split_to_dummy(x, sep = NULL)
}
\arguments{
\item{x}{var to be split and converted to dummy representation}

\item{sep}{separator for strsplit function, if sep=NULL then splitting is ommited}
}
\value{
matrix of dummy vars
}
\description{
data canversion to numeric, dummy variables for multi-level categorical vars
}
