% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.to_tbl_cat}
\alias{.to_tbl_cat}
\title{Create table of abs and rel frequencies (%) for categorical variables}
\usage{
.to_tbl_cat(pv)
}
\arguments{
\item{list}{of two: 1: p values semgentation levels, 2: contigency table (abs, %)}
}
\value{
table ready to export
}
\description{
Create table of abs and rel frequencies (%) for categorical variables
}
