% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{.to_tbl_num}
\alias{.to_tbl_num}
\title{Create table of summary stats for continuous variables}
\usage{
.to_tbl_num(pv)
}
\arguments{
\item{list}{of two: 1: p values semgentation levels, 2: quantiles for segm. levels}
}
\value{
table ready to export
}
\description{
Create table of summary stats for continuous variables
}
