% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{es_cdelta}
\alias{es_cdelta}
\title{Pairwise effect size - Cliff's delta}
\usage{
es_cdelta(ref, comp)
}
\arguments{
\item{ref}{num array, variable values for group 1 (reference group)}

\item{comp}{ref num array, variable values for group 2 (comparison group)}
}
\value{
Cliff's delta value
}
\description{
Pairwise effect size - Cliff's delta
}
