% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{infer_dtype}
\alias{infer_dtype}
\title{infer variable data type}
\usage{
infer_dtype(x, gsub_pattern = "<|>", n_un_lim = 5, n_un_perc_lim = 10)
}
\arguments{
\item{x}{variable of interest}

\item{gsub_pattern}{regex for str to remove (e.g., '<|>')}

\item{n_un_lim}{number of unique values in x to define numeric (see details)}

\item{n_un_perc_lim}{fraction of unique values in x to define numeric (see details)}
}
\value{
data type as string
}
\description{
infer variable data type
}
\details{
dtype inferred is one of num, str, logic, x should be free of na or non-informative values (see fct .rm_na). Parameters n_un_lim and n_un_frac_lim define absolute and fraction of unique values, respectively, to define x as numeric. This is to exclude categorical values from being defined as numeric (e.g., x levels of low medium high).
}
