% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts_cleanup.R
\name{infer_level}
\alias{infer_level}
\title{infer if variable is of type multi-response}
\usage{
infer_level(x, split_cutoff = 20, n_un, sep = c(",", ";"))
}
\arguments{
\item{x}{variable to be cleaned}

\item{split_cutoff}{cut-off in percent for minimal entries with multi-responses}

\item{n_un}{length unique values of x}

\item{sep}{separators}
}
\value{
list of possible separators
}
\description{
infer if variable is of type multi-response
}
\details{
Multi-response variables defined as strings with where multiple values are separated with comma, semicolon or any other alpha-numeric value indicated in sep argument
}
