% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calcStat.R, R/prep_var.R
\name{prep_var}
\alias{prep_var}
\title{Summaris variable dependent on stratification level}
\usage{
prep_var(
  x,
  y,
  x.multi = F,
  y.multi = F,
  x.dtype = c("num", "str", "fac"),
  x.levs = NULL,
  y.dtype = c("num", "str", "fac"),
  y.levs = NULL,
  n_perc = 5
)

prep_var(
  x,
  y,
  x.multi = F,
  y.multi = F,
  x.dtype = c("num", "str", "fac"),
  x.levs = NULL,
  y.dtype = c("num", "str", "fac"),
  y.levs = NULL,
  n_perc = 5
)
}
\arguments{
\item{list}{of two: 1: p values semgentation levels, 2: contigency table (abs, %)}
}
\value{
table ready to export

table ready to export
}
\description{
Summaris variable dependent on stratification level

Summaris variable dependent on stratification level
}
