% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_var.R
\name{prep_var}
\alias{prep_var}
\title{Variable summary on stratification level}
\usage{
prep_var(
  x,
  y,
  x.multi = F,
  y.multi = F,
  x.dtype = c("num", "str", "fac"),
  x.levs = NULL,
  y.dtype = c("num", "str", "fac"),
  y.levs = NULL,
  n_perc = 5
)
}
\arguments{
\item{x}{variable for summary stats}

\item{y}{stratification variable}

\item{x.multi}{logic, indicating if x is mult-response variable variable to be cleaned}

\item{y.multi}{logic, indicating if y is mult-response variable variable to be cleaned}

\item{x.dtype}{data type of x ('num', 'str', 'fac', logic)}

\item{x.levs}{levels of x in case x is factor}

\item{y.dtype}{data type of y ('num', 'str', 'fac', logic)}

\item{y.levs}{levels of y in case y is factor}

\item{n_perc}{frequency cutoff in percentage to combine categorical levels into class 'other'}
}
\value{
table ready to export
}
\description{
Variable summary on stratification level
}
